requirements.txt
Authlib==1.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
idna==3.4
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
validators==0.22.0
weaviate-client==3.25.2
